% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_functions.R
\name{construct_symmetric_matrix}
\alias{construct_symmetric_matrix}
\title{Construct a Symmetric Matrix from a Vector, typicaly, we use this to contruct the variogram matrix based on the upper triangular entries.}
\usage{
construct_symmetric_matrix(vec)
}
\arguments{
\item{vec}{A numeric vector of length \eqn{d(d - 1)/2}, representing the upper triangle (excluding the diagonal) of a \eqn{d \times d} symmetric matrix.}
}
\value{
A \eqn{d \times d} symmetric matrix with ones on the diagonal and the upper/lower triangular entries filled from \code{vec}.
}
\description{
Converts a vector containing the upper triangular entries (excluding the diagonal) of a symmetric matrix into the full symmetric matrix with ones on the diagonal.
}
