% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_functions.R
\name{EDS}
\alias{EDS}
\title{Compute the extreme directions score}
\usage{
EDS(B, A)
}
\arguments{
\item{B}{A numeric matrix.}

\item{A}{A numeric matrix with the same number of columns as \code{B}.}
}
\value{
A numeric value between 0 and 1 representing the proportion of missidentified signatures (or equivalently extreme directions); See Equation (4.3) from Mourahib, Kiriliouk, and Segers (2025).
}
\description{
Calculates a Jaccard distance between matrices \code{B} and \code{A} based on the proportion of the non-similar signatures
}
\references{
Mourahib, A., Kiriliouk, A., & Segers, J. (2025). A penalized least squares estimator for extreme-value mixture models. \emph{arXiv preprint} arXiv:2506.15272.
}
