% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_functions.R
\name{shuffleCols}
\alias{shuffleCols}
\title{Reorder Columns to Match a Target Matrix}
\usage{
shuffleCols(start, A)
}
\arguments{
\item{start}{A numeric matrix with \code{d} rows. Typically an estimated coefficient matrix.}

\item{A}{A numeric matrix representing the true coefficient matrix.}
}
\value{
A reordered version of \code{start}, where the first \code{ncol(A)} columns are matched to \code{A} to minimize squared differences, and remaining columns are placed at the end.
}
\description{
Reorders the columns of the input matrix \code{start} to best match the columns of the matrix \code{A}, based on minimizing the squared Euclidean distance without normalization, as in Equation (4.5) from Mourahib, Kiriliouk, and Segers (2025).
}
\references{
Mourahib, A., Kiriliouk, A., & Segers, J. (2025). A penalized least squares estimator for extreme-value mixture models. \emph{arXiv preprint} arXiv:2506.15272.
}
