% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_functions.R
\name{SMSE_HR}
\alias{SMSE_HR}
\title{Compute standardized Mean Squared Error in Equation (4.4) from Mourahib, Kiriliouk, and Segers (2025) for the mixture Hüsler–Reiss model}
\usage{
SMSE_HR(list_matrices, A, ind = NULL)
}
\arguments{
\item{list_matrices}{A list where each element contains an \code{Estimation} list with \code{pls_matrix} and \code{pls_dep} components.}

\item{A}{The true coefficient matrix.}

\item{ind}{Optional integer vector of indices to exclude when comparing dependence parameters.}
}
\value{
A numeric value representing the aggregated standardized mean squared error over all estimates.
}
\description{
Calculates a standardized mean squared error (SMSE) between a list of estimated matrices and a true coefficient matrix,
incorporating both matrix and dependence parameter errors. Scaling adjusts for variability across estimates.
}
\references{
Mourahib, A., Kiriliouk, A., & Segers, J. (2025). A penalized least squares estimator for extreme-value mixture models. \emph{arXiv preprint} arXiv:2506.15272.
}
