% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_functions.R
\name{SMSE_log}
\alias{SMSE_log}
\title{Standardized mean squarred error in Equation (4.4) from Mourahib, Kiriliouk, and Segers (2025) for the mixture logistic model}
\usage{
SMSE_log(list, A, dep)
}
\arguments{
\item{list}{A list of estimation results, where each element is expected to contain a component \code{Estimation$matrix} (estimated matrix)
and \code{Estimation$dep} (estimated dependence parameter).}

\item{A}{The true coefficient matrix (numeric matrix) with dimensions \code{d} by \code{r}.}

\item{dep}{The true dependence parameter (numeric scalar).}
}
\value{
A numeric scalar representing the standardized mean squared error combining the matrix and dependence parameter estimation errors.
}
\description{
Computes a standardized mean squared error (SMSE) between a list of estimated coefficient matrices and a true coefficient matrix,
adjusted for differences in the number of columns. It also incorporates the squared error of an associated dependence parameter.
}
\details{
The function first ensures each estimated matrix has the same number of columns as the true matrix by padding with zeros if necessary.
It then calculates the average normalized squared differences between the estimated and true matrices, scaling by the maximum estimated values across seeds to avoid division by zero.
The dependence parameter error is incorporated as a standardized squared difference.
}
\references{
Mourahib, A., Kiriliouk, A., & Segers, J. (2025). A penalized least squares estimator for extreme-value mixture models. \emph{arXiv preprint} arXiv:2506.15272.
}
